backend/requirements.txt
email-validator==2.3.0
fastapi==0.135.3
faster-whisper==1.2.1
httpx==0.28.1
openai==2.11.0
bcrypt>=4.2,<5
python-jose[cryptography]==3.5.0
sqlalchemy==2.0.45
stripe>=12.5,<16
opencc-python-reimplemented==0.1.7
pytest==9.0.3
uvicorn[standard]==0.44.0
yt-dlp==2026.3.17
